moving log_prob and sample into generic SSM base class.
